Upload script for cw20-base and cw20-ics20
